azure_functions/requirements.txt
azure-functions
azure-storage-blob
requests==2.25.1
beautifulsoup4==4.9.3
colorama==0.4.4
applicationinsights==0.11.10
